Reinforcement learning with Q-Learning v.00004
The result is collect now.
But, in face this sample is not a suitable sample for Q-Learning. First
, it does use the Bellman Equation. Second , the state structure has
only one layer. So, this sample  just has the same result of Contextual
Bandits .
But the code has  provide the ability to build a multilayer
structure，and it also show how to add state dynamically. And it show the
rounded system of a Q Learning system too.
Now, i am still not much satisfied with current code, and i feel it's
still too complex and not limpid. I will upset event when i am
debugging.
So, i will adjust the code fisrt. And maybe provide a sample of
wayfinding in a gridworld Laster， just as unity do